requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-migrate==4.0.4
flask-bcrypt==1.0.1
flask-cors==3.0.10
flask-wtf==1.1.1
email-validator==2.0.0
werkzeug==2.3.4
wtforms==3.0.1
itsdangerous==2.1.2
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.31.0
pillow==9.5.0
sqlalchemy==2.0.15
pymysql==1.0.3
cryptography==41.0.1